Atualização do projeto

+++ DirectAlertBot/Services/AlertService.cs
@@ -1,5 +1,7 @@
 ﻿using DirectAlertBot.Data;
-using DirectAlertBot.Models;
+using System.Linq;
+using DirectAlertBot.Interfaces;
+using DirectAlertBot.Entities;
 
 namespace DirectAlertBot.Services
 {
@@ -12,21 +14,40 @@ namespace DirectAlertBot.Services
             _context = context;
         }
 
-        public void Insert(Alert alert)
+        public void InsertAlert(Alert alert)
         {
             _context.Alerts.Add(alert);
             _context.SaveChanges();
         }
 
-        public Alert FindById(int id)
+        public Alert? FindAlertById(int id)
         {
             return _context.Alerts.Find(id);
         }
 
-        public void Remove(int id)
+        public void UpdateAlert(Alert entity)
         {
-            Alert obj = FindById(id);
+            _context.Alerts.Update(entity);
+            _context.SaveChanges();
+        }
+
+        public void RemoveAlert(int id)
+        {
+            Alert? obj = FindAlertById(id);
+            if (obj == null) 
+                return;
             _context.Alerts.Remove(obj);
+            _context.SaveChanges();
+        }
+
+        public Client? FindClientByUserId(long userId)
+        {
+            return _context.Clients.FirstOrDefault(x => x.UserId == userId);
+        }
+
+        public void InsertClient(Client client)
+        {
+            _context.Clients.Add(client);
             _context.SaveChanges();
         }
     }
